+++ SuperStore P3/EcoPower_Logistics.Data/Repositories/IGenericRepository.cs
@@ -5,7 +5,7 @@ namespace EcoPower_Logistics.Data.Repositories
     public interface IGenericRepository<T> where T : class
     {
         IEnumerable<T> GetAll();
-        T GetById(short id);
+        T GetById(int id);
         void Add(T entity);
         void Remove(T entity);
         void Update(T entity);
