+++ SuperStore P3/EcoPower_Logistics.Data/Repositories/IOrderDetailsRepository.cs
@@ -4,7 +4,7 @@ namespace EcoPower_Logistics.Data.Repositories
 {
     public interface IOrderDetailsRepository : IGenericRepository<OrderDetail>
     {
-        OrderDetail GetOrderDetailById(short? id);
+        OrderDetail GetOrderDetailById(int? id);
         IEnumerable<OrderDetail> GetAllOrderDetails();
         void AddOrderDetail(OrderDetail entity);
         void UpdateOrderDetail(OrderDetail entity);
