+++ SuperStore P3/EcoPower_Logistics.Data/Repositories/IProductRepository.cs
@@ -4,7 +4,7 @@ namespace EcoPower_Logistics.Data.Repositories
 {
     public interface IProductRepository : IGenericRepository<Product>
     {
-        Product GetProductById(short? id);
+        Product GetProductById(int? id);
         IEnumerable<Product> GetAllProducts();
         void AddProduct(Product entity);
         void UpdateProduct(Product entity);
