+++ SuperStore P3/EcoPower_Logistics.Service/Services/ICustomerService.cs
@@ -4,7 +4,7 @@ namespace EcoPower_Logistics.Service.Services
 {
     public interface ICustomerService
     {
-        Customer GetCustomerById(short? id);
+        Customer GetCustomerById(int? id);
         IEnumerable<Customer> GetAllCustomers();
         void AddCustomer(Customer entity);
         void UpdateCustomer(Customer entity);
