+++ SuperStore P3/EcoPower_Logistics.Service/Services/IOrderDetailsService.cs
@@ -4,7 +4,7 @@ namespace EcoPower_Logistics.Service.Services
 {
     public interface IOrderDetailsService
     {
-        OrderDetail GetOrderDetailById(short? id);
+        OrderDetail GetOrderDetailById(int? id);
         IEnumerable<OrderDetail> GetAllOrderDetails();
         void AddOrderDetail(OrderDetail entity);
         void UpdateOrderDetail(OrderDetail entity);
