+++ SuperStore P3/EcoPower_Logistics.Service/Services/IOrderService.cs
@@ -4,7 +4,7 @@ namespace EcoPower_Logistics.Service.Services
 {
     public interface IOrderService
     {
-        Order GetOrderById(short? id);
+        Order GetOrderById(int? id);
         IEnumerable<Order> GetAllOrders();
         void AddOrder(Order entity);
         void UpdateOrder(Order entity);
