+++ SuperStore P3/EcoPower_Logistics.Service/Services/IProductService.cs
@@ -4,7 +4,7 @@ namespace EcoPower_Logistics.Service.Services
 {
     public interface IProductService
     {
-        Product GetProductById(short? id);
+        Product GetProductById(int? id);
         IEnumerable<Product> GetAllProducts();
         void AddProduct(Product entity);
         void UpdateProduct(Product entity);
